+++ index.js
@@ -1,50 +1,31 @@
 'use strict';
 
-var vm = require('vm');
-var requestModule = require('request-promise');
-var errors = require('./errors');
-var decodeEmails = require('./lib/email-decode.js');
+const requestModule = require('request-promise');
+const sandbox = require('./lib/sandbox');
+const decodeEmails = require('./lib/email-decode.js');
+const getDefaultHeaders = require('./lib/headers');
+const brotli = require('./lib/brotli');
 
-var USER_AGENTS = [
-  'Ubuntu Chromium/34.0.1847.116 Chrome/34.0.1847.116 Safari/537.36',
-  'Mozilla/5.0 (X11; Linux x86_64) AppleWebKit/537.21 (KHTML, like Gecko) konqueror/4.14.10 Safari/537.21',
-  'Mozilla/5.0 (iPad; CPU OS 5_1 like Mac OS X) AppleWebKit/534.46 (KHTML, like Gecko ) Version/5.1 Mobile/9B176 Safari/7534.48.3',
-  'Mozilla/5.0 (iPad; U; CPU OS 3_2 like Mac OS X; en-us) AppleWebKit/531.21.10 (KHTML, like Gecko) Version/4.0.4 Mobile/7B334b Safari/531.21.10',
-  'Mozilla/5.0 (compatible; MSIE 9.0; Windows NT 6.1; WOW64; Trident/5.0; SLCC2; Media Center PC 6.0; InfoPath.3; MS-RTC LM 8; Zune 4.7)',
-  'Mozilla/5.0 (Windows Phone 8.1; ARM; Trident/7.0; Touch; rv:11.0; IEMobile/11.0; NOKIA; Lumia 630) like Gecko',
-  'Mozilla/5.0 (compatible; MSIE 10.0; Windows Phone 8.0; Trident/6.0; IEMobile/10.0; ARM; Touch; NOKIA; Lumia 920)',
-  'Mozilla/5.0 (Linux; U; Android 2.2; en-us; Sprint APA9292KT Build/FRF91) AppleWebKit/533.1 (KHTML, like Gecko) Version/4.0 Mobile Safari/533.1',
-  'Mozilla/5.0 (X11; Linux x86_64; rv:2.2a1pre) Gecko/20100101 Firefox/4.2a1pre',
-  'Mozilla/5.0 (SymbianOS/9.1; U; en-us) AppleWebKit/413 (KHTML, like Gecko) Safari/413 es65',
-  'Mozilla/5.0 (Linux; Android 6.0; Nexus 5X Build/MDB08L) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/53.0.2785.124 Mobile Safari/537.36',
-  'Mozilla/5.0 (X11; U; FreeBSD i386; de-CH; rv:1.9.2.8) Gecko/20100729 Firefox/3.6.8'
-];
+const {
+  RequestError,
+  CaptchaError,
+  CloudflareError,
+  ParserError
+} = require('./errors');
 
-var DEFAULT_USER_AGENT = randomUA();
-
-var VM_OPTIONS = {
-  contextOrigin: 'cloudflare:challenge.js',
-  contextCodeGeneration: { strings: true, wasm: false },
-  timeout: 5000
-};
+const HOST = Symbol('host');
 
 module.exports = defaults.call(requestModule);
 
 function defaults (params) {
   // isCloudScraper === !isRequestModule
-  var isRequestModule = this === requestModule;
+  const isRequestModule = this === requestModule;
 
-  var defaultParams = (!isRequestModule && this.defaultParams) || {
+  let defaultParams = (!isRequestModule && this.defaultParams) || {
     requester: requestModule,
     // Cookies should be enabled
     jar: requestModule.jar(),
-    headers: {
-      'Connection': 'keep-alive',
-      'User-Agent': DEFAULT_USER_AGENT,
-      'Cache-Control': 'private',
-      'Accept': 'application/xml,application/xhtml+xml,text/html;q=0.9, text/plain;q=0.8,image/png,*/*;q=0.5',
-      'Accept-Language': 'en-US,en;q=0.9'
-    },
+    headers: getDefaultHeaders({ 'Host': HOST }),
     // Reduce Cloudflare's timeout to cloudflareMaxTimeout if it is excessive
     cloudflareMaxTimeout: 30000,
     // followAllRedirects - follow non-GET HTTP 3xx responses as redirects
@@ -52,14 +33,17 @@ function defaults (params) {
     // Support only this max challenges in row. If CF returns more, throw an error
     challengesToSolve: 3,
     // Remove Cloudflare's email protection
-    decodeEmails: false
+    decodeEmails: false,
+    // Support gzip encoded responses
+    gzip: true
   };
 
   // Object.assign requires at least nodejs v4, request only test/supports v6+
   defaultParams = Object.assign({}, defaultParams, params);
 
-  var cloudscraper = requestModule.defaults
+  const cloudscraper = requestModule.defaults
     .call(this, defaultParams, function (options) {
+      validateRequest(options);
       return performRequest(options, true);
     });
 
@@ -79,9 +63,7 @@ function defaults (params) {
   return cloudscraper;
 }
 
-// This function is wrapped to ensure that we get new options on first call.
-// The options object is reused in subsequent calls when calling it directly.
-function performRequest (options, isFirstRequest) {
+function validateRequest (options) {
   // Prevent overwriting realEncoding in subsequent calls
   if (!('realEncoding' in options)) {
     // Can't just do the normal options.encoding || 'utf8'
@@ -105,15 +87,26 @@ function performRequest (options, isFirstRequest) {
       'got ' + typeof (options.cloudflareMaxTimeout) + ' instead.');
   }
 
-  // This should be the default export of either request or request-promise.
-  var requester = options.requester;
-
-  if (typeof requester !== 'function') {
+  if (typeof options.requester !== 'function') {
     throw new TypeError('Expected `requester` option to be a function, got ' +
-        typeof (requester) + ' instead.');
+      typeof (options.requester) + ' instead.');
   }
+}
 
-  var request = requester(options);
+// This function is wrapped to ensure that we get new options on first call.
+// The options object is reused in subsequent calls when calling it directly.
+function performRequest (options, isFirstRequest) {
+  // This should be the default export of either request or request-promise.
+  const requester = options.requester;
+
+  // Note that request is always an instanceof ReadableStream, EventEmitter
+  // If the requester is request-promise, it is also thenable.
+  const request = requester(options);
+
+  // We must define the host header ourselves to preserve case and order.
+  if (request.getHeader('host') === HOST) {
+    request.setHeader('host', request.uri.host);
+  }
 
   // If the requester is not request-promise, ensure we get a callback.
   if (typeof request.callback !== 'function') {
@@ -139,7 +132,7 @@ function performRequest (options, isFirstRequest) {
       onRequestResponse(options, null, response, body);
     });
 
-  // Indicate that this is a cloudscraper request, required by test/helper.
+  // Indicate that this is a cloudscraper request
   request.cloudscraper = true;
   return request;
 }
@@ -147,12 +140,12 @@ function performRequest (options, isFirstRequest) {
 // The argument convention is options first where possible, options
 // always before response, and body always after response.
 function onRequestResponse (options, error, response, body) {
-  var callback = options.callback;
+  const callback = options.callback;
 
   // Encoding is null so body should be a buffer object
   if (error || !body || !body.toString) {
     // Pure request error (bad connection, wrong url, etc)
-    return callback(new errors.RequestError(error, options, response));
+    return callback(new RequestError(error, options, response));
   }
 
   response.responseStartTime = Date.now();
@@ -164,23 +157,33 @@ function onRequestResponse (options, error, response, body) {
     return callback(null, response, body);
   }
 
+  // Decompress brotli compressed responses
+  if (/\bbr\b/i.test('' + response.caseless.get('content-encoding'))) {
+    if (!brotli.isAvailable) {
+      const cause = 'Received a Brotli compressed response. Please install brotli';
+      return callback(new RequestError(cause, options, response));
+    }
+
+    response.body = body = brotli.decompress(body);
+  }
+
   if (response.isCloudflare && response.isHTML) {
     onCloudflareResponse(options, response, body);
   } else {
-    processResponseBody(options, response, body);
+    onRequestComplete(options, response, body);
   }
 }
 
 function onCloudflareResponse (options, response, body) {
-  var callback = options.callback;
+  const callback = options.callback;
 
-  var stringBody;
-  var isChallenge;
-  var isRedirectChallenge;
+  let stringBody;
+  let isChallenge;
+  let isRedirectChallenge;
 
   if (body.length < 1) {
     // This is a 4xx-5xx Cloudflare response with an empty body.
-    return callback(new errors.CloudflareError(response.statusCode, options, response));
+    return callback(new CloudflareError(response.statusCode, options, response));
   }
 
   stringBody = body.toString('utf8');
@@ -188,104 +191,106 @@ function onCloudflareResponse (options, response, body) {
   try {
     validate(options, response, stringBody);
   } catch (error) {
+    if (error instanceof CaptchaError && typeof options.onCaptcha === 'function') {
+      // Give users a chance to solve the reCAPTCHA via services such as anti-captcha.com
+      return onCaptcha(options, response, stringBody);
+    }
+
     return callback(error);
   }
 
   isChallenge = stringBody.indexOf('a = document.getElementById(\'jschl-answer\');') !== -1;
 
   if (isChallenge) {
-    return solveChallenge(options, response, stringBody);
+    return onChallenge(options, response, stringBody);
   }
 
   isRedirectChallenge = stringBody.indexOf('You are being redirected') !== -1 ||
     stringBody.indexOf('sucuri_cloudproxy_js') !== -1;
 
   if (isRedirectChallenge) {
-    return setCookieAndReload(options, response, stringBody);
+    return onRedirectChallenge(options, response, stringBody);
   }
 
   // 503 status is always a challenge
   if (response.statusCode === 503) {
-    return solveChallenge(options, response, stringBody);
+    return onChallenge(options, response, stringBody);
   }
 
   // All is good
-  processResponseBody(options, response, body);
+  onRequestComplete(options, response, body);
 }
 
 function validate (options, response, body) {
-  var match;
+  let match;
 
   // Finding captcha
   if (body.indexOf('why_captcha') !== -1 || /cdn-cgi\/l\/chk_captcha/i.test(body)) {
-    throw new errors.CaptchaError('captcha', options, response);
+    // Convenience boolean
+    response.isCaptcha = true;
+    throw new CaptchaError('captcha', options, response);
   }
 
   // Trying to find '<span class="cf-error-code">1006</span>'
   match = body.match(/<\w+\s+class="cf-error-code">(.*)<\/\w+>/i);
 
   if (match) {
-    var code = parseInt(match[1]);
-    throw new errors.CloudflareError(code, options, response);
+    let code = parseInt(match[1]);
+    throw new CloudflareError(code, options, response);
   }
 
   return false;
 }
 
-function solveChallenge (options, response, body) {
-  var callback = options.callback;
-  var cause;
-  var error;
+function onChallenge (options, response, body) {
+  const callback = options.callback;
+  const uri = response.request.uri;
+  // The query string to send back to Cloudflare
+  const payload = { /* s, jschl_vc, pass, jschl_answer */ };
+
+  let cause;
+  let error;
 
   if (options.challengesToSolve === 0) {
     cause = 'Cloudflare challenge loop';
-    error = new errors.CloudflareError(cause, options, response);
+    error = new CloudflareError(cause, options, response);
     error.errorType = 4;
 
     return callback(error);
   }
 
-  var timeout = parseInt(options.cloudflareTimeout);
-  var uri = response.request.uri;
-  // The query string to send back to Cloudflare
-  // var payload = { s, jschl_vc, pass, jschl_answer };
-  var payload = {};
-  var sandbox;
-  var match;
+  let timeout = parseInt(options.cloudflareTimeout);
+  let match;
 
   match = body.match(/name="s" value="(.+?)"/);
-
   if (match) {
     payload.s = match[1];
   }
 
   match = body.match(/name="jschl_vc" value="(\w+)"/);
-
   if (!match) {
     cause = 'challengeId (jschl_vc) extraction failed';
-    return callback(new errors.ParserError(cause, options, response));
+    return callback(new ParserError(cause, options, response));
   }
 
   payload.jschl_vc = match[1];
 
   match = body.match(/name="pass" value="(.+?)"/);
-
   if (!match) {
     cause = 'Attribute (pass) value extraction failed';
-    return callback(new errors.ParserError(cause, options, response));
+    return callback(new ParserError(cause, options, response));
   }
 
   payload.pass = match[1];
 
   match = body.match(/getElementById\('cf-content'\)[\s\S]+?setTimeout.+?\r?\n([\s\S]+?a\.value\s*=.+?)\r?\n(?:[^{<>]*},\s*(\d{4,}))?/);
-
   if (!match) {
     cause = 'setTimeout callback extraction failed';
-    return callback(new errors.ParserError(cause, options, response));
+    return callback(new ParserError(cause, options, response));
   }
 
   if (isNaN(timeout)) {
-    if (match.length > 2) {
+    if (match[2] !== undefined) {
       timeout = parseInt(match[2]);
 
       if (timeout > options.cloudflareMaxTimeout) {
@@ -297,7 +302,7 @@ function solveChallenge (options, response, body) {
       }
     } else {
       cause = 'Failed to parse challenge timeout';
-      return callback(new errors.ParserError(cause, options, response));
+      return callback(new ParserError(cause, options, response));
     }
   }
 
@@ -305,16 +310,16 @@ function solveChallenge (options, response, body) {
   response.challenge = match[1] + '; a.value';
 
   try {
-    sandbox = createSandbox({ uri: uri, body: body });
-    payload.jschl_answer = vm.runInNewContext(response.challenge, sandbox, VM_OPTIONS);
+    const ctx = new sandbox.Context({ hostname: uri.hostname, body });
+    payload.jschl_answer = sandbox.eval(response.challenge, ctx);
   } catch (error) {
     error.message = 'Challenge evaluation failed: ' + error.message;
-    return callback(new errors.ParserError(error, options, response));
+    return callback(new ParserError(error, options, response));
   }
 
   if (isNaN(payload.jschl_answer)) {
     cause = 'Challenge answer is not a number';
-    return callback(new errors.ParserError(cause, options, response));
+    return callback(new ParserError(cause, options, response));
   }
 
   // Prevent reusing the headers object to simplify unit testing.
@@ -328,44 +333,151 @@ function solveChallenge (options, response, body) {
   }
   // Set the query string and decrement the number of challenges to solve.
   options.qs = payload;
-  options.challengesToSolve = options.challengesToSolve - 1;
+  options.challengesToSolve -= 1;
 
   // Make request with answer after delay.
   timeout -= Date.now() - response.responseStartTime;
   setTimeout(performRequest, timeout, options, false);
 }
 
-function setCookieAndReload (options, response, body) {
-  var callback = options.callback;
+// Parses the reCAPTCHA form and hands control over to the user
+function onCaptcha (options, response, body) {
+  const callback = options.callback;
+  // UDF that has the responsibility of returning control back to cloudscraper
+  const handler = options.onCaptcha;
+  // The form data to send back to Cloudflare
+  const payload = { /* s, g-re-captcha-response */ };
 
-  var match = body.match(/S='([^']+)'/);
+  let cause;
+  let match;
 
+  match = body.match(/<form(?: [^<>]*)? id=["']?challenge-form['"]?(?: [^<>]*)?>([\S\s]*?)<\/form>/);
   if (!match) {
-    var cause = 'Cookie code extraction failed';
-    return callback(new errors.ParserError(cause, options, response));
+    cause = 'Challenge form extraction failed';
+    return callback(new ParserError(cause, options, response));
   }
 
-  var base64EncodedCode = match[1];
-  response.challenge = Buffer.from(base64EncodedCode, 'base64').toString('ascii');
+  // Defining response.challengeForm for debugging purposes
+  const form = response.challengeForm = match[1];
 
-  try {
-    var sandbox = createSandbox({});
-    // Evaluate cookie setting code
-    vm.runInNewContext(response.challenge, sandbox, VM_OPTIONS);
-
-    options.jar.setCookie(sandbox.document.cookie, response.request.uri.href, { ignoreError: true });
-  } catch (error) {
-    error.message = 'Cookie code evaluation failed: ' + error.message;
-    return callback(new errors.ParserError(error, options, response));
+  match = form.match(/\/recaptcha\/api\/fallback\?k=([^\s"'<>]*)/);
+  if (!match) {
+    // The site key wasn't inside the form so search the entire document
+    match = body.match(/data-sitekey=["']?([^\s"'<>]*)/);
+    if (!match) {
+      cause = 'Unable to find the reCAPTCHA site key';
+      return callback(new ParserError(cause, options, response));
+    }
   }
 
-  options.challengesToSolve = options.challengesToSolve - 1;
+  // Everything that is needed to solve the reCAPTCHA
+  response.captcha = { siteKey: match[1], form: payload };
+
+  // Adding formData
+  match = form.match(/<input(?: [^<>]*)? name=[^<>]+>/g);
+  if (!match) {
+    cause = 'Challenge form is missing inputs';
+    return callback(new ParserError(cause, options, response));
+  }
+
+  const inputs = match;
+  // Only adding inputs that have both a name and value defined
+  for (let name, value, i = 0; i < inputs.length; i++) {
+    name = inputs[i].match(/name=["']?([^\s"'<>]*)/);
+    if (name) {
+      value = inputs[i].match(/value=["']?([^\s"'<>]*)/);
+      if (value) {
+        payload[name[1]] = value[1];
+      }
+    }
+  }
+
+  // Sanity check
+  if (!payload['s']) {
+    cause = 'Challenge form is missing secret input';
+    return callback(new ParserError(cause, options, response));
+  }
+
+  // The callback used to green light form submission
+  const submit = function (error) {
+    if (error) {
+      // Pass an user defined error back to the original request call
+      return callback(new CaptchaError(error, options, response));
+    }
+
+    onSubmitCaptcha(options, response, body);
+  };
+
+  // This seems like an okay-ish API (fewer arguments to the handler)
+  response.captcha.submit = submit;
+
+  // We're handing control over to the user now.
+  const thenable = handler(options, response, body);
+  // Handle the case where the user returns a promise
+  if (thenable && typeof thenable.then === 'function') {
+    // eslint-disable-next-line promise/catch-or-return
+    thenable.then(submit, function (error) {
+      if (!error) {
+        // The user broke their promise with a falsy error
+        submit(new Error('Falsy error'));
+      } else {
+        submit(error);
+      }
+    });
+  }
+}
+
+function onSubmitCaptcha (options, response) {
+  const callback = options.callback;
+  const uri = response.request.uri;
+
+  if (!response.captcha.form['g-recaptcha-response']) {
+    const cause = 'Form submission without g-recaptcha-response';
+    return callback(new CaptchaError(cause, options, response));
+  }
+
+  options.method = 'GET';
+  options.qs = response.captcha.form;
+  // Prevent reusing the headers object to simplify unit testing.
+  options.headers = Object.assign({}, options.headers);
+  // Use the original uri as the referer and to construct the form action.
+  options.headers['Referer'] = uri.href;
+  options.uri = uri.protocol + '//' + uri.host + '/cdn-cgi/l/chk_captcha';
 
   performRequest(options, false);
 }
 
-function processResponseBody (options, response, body) {
-  var callback = options.callback;
+function onRedirectChallenge (options, response, body) {
+  const callback = options.callback;
+  const uri = response.request.uri;
+
+  const match = body.match(/S='([^']+)'/);
+  if (!match) {
+    const cause = 'Cookie code extraction failed';
+    return callback(new ParserError(cause, options, response));
+  }
+
+  const base64EncodedCode = match[1];
+  response.challenge = Buffer.from(base64EncodedCode, 'base64').toString('ascii');
+
+  try {
+    // Evaluate cookie setting code
+    const ctx = new sandbox.Context();
+    sandbox.eval(response.challenge, ctx);
+
+    options.jar.setCookie(ctx.document.cookie, uri.href, { ignoreError: true });
+  } catch (error) {
+    error.message = 'Cookie code evaluation failed: ' + error.message;
+    return callback(new ParserError(error, options, response));
+  }
+
+  options.challengesToSolve -= 1;
+
+  performRequest(options, false);
+}
+
+function onRequestComplete (options, response, body) {
+  const callback = options.callback;
 
   if (typeof options.realEncoding === 'string') {
     body = body.toString(options.realEncoding);
@@ -380,48 +492,4 @@ function processResponseBody (options, response, body) {
   }
 
   callback(null, response, body);
-}
-
-function createSandbox (options) {
-  if (options.body) {
-    var body = options.body;
-    var href = 'http://' + options.uri.hostname + '/';
-    var cache = Object.create(null);
-    var keys = [];
-
-    // Sandbox for standard IUAM JS challenge
-    return Object.assign({
-      atob: function (str) {
-        return Buffer.from(str, 'base64').toString('binary');
-      },
-      document: {
-        createElement: function () {
-          return { firstChild: { href: href } };
-        },
-        getElementById: function (id) {
-          if (keys.indexOf(id) === -1) {
-            var re = new RegExp(' id=[\'"]?' + id + '[^>]*>([^<]+)');
-            var match = body.match(re);
-
-            keys.push(id);
-            cache[id] = match === null ? match : { innerHTML: match[1] };
-          }
-
-          return cache[id];
-        }
-      }
-    }, options.context);
-  }
-
-  // Sandbox used in setCookieAndReload
-  return Object.assign({
-    location: {
-      reload: function () {}
-    },
-    document: {}
-  }, options.context);
-}
-
-function randomUA () {
-  return USER_AGENTS[Math.floor(Math.random() * USER_AGENTS.length)];
 }